+++ output/builtin.go
@@ -9,8 +9,6 @@ import (
 	"log"
 	"os"
 
-	"github.com/thalesfsp/sypl/elasticsearch"
-	"github.com/thalesfsp/sypl/formatter"
 	"github.com/thalesfsp/sypl/level"
 	"github.com/thalesfsp/sypl/processor"
 	"github.com/thalesfsp/sypl/safebuffer"
@@ -82,99 +80,4 @@ func SafeBuffer(maxLevel level.Level, processors ...processor.IProcessor) (*safe
 	o := New("Buffer", maxLevel, &buf, processors...)
 
 	return &buf, o
-}
-
-// ElasticSearchConfig is the ElasticSearch configuration.
-type ElasticSearchConfig = elasticsearch.Config
-
-// ElasticSearchDynamicIndexFunc is the name of the index to be used, evaluated
-// at the insertion time.
-type ElasticSearchDynamicIndexFunc = elasticsearch.DynamicIndexFunc
-
-// ElasticSearch is a built-in `output` - named `ElasticSearch`, that writes to
-// ElasticSearch.
-//
-// NOTE: By default, data is JSON-formatted.
-// NOTE: `DocumentID` is automatically generated.
-// NOTE: It's the caller's responsibility to create the index, define its
-// mapping, and settings.
-func ElasticSearch(
-	indexName string,
-	esConfig ElasticSearchConfig,
-	maxLevel level.Level,
-	processors ...processor.IProcessor,
-) IOutput {
-	return ElasticSearchWithDynamicIndex(
-		func() string { return indexName },
-		esConfig,
-		maxLevel,
-		processors...,
-	)
-}
-
-// ElasticSearchTagMapItem X.
-type ElasticSearchTagMapItem struct {
-	DynamicIndexFunc ElasticSearchDynamicIndexFunc
-	Level            level.Level
-}
-
-// ElasticSearchTagMap is a map of tags to index names (DynamicIndexFunc).
-type ElasticSearchTagMap = map[string]ElasticSearchTagMapItem
-
-// NewElasticSearchTagMapItem X.
-func NewElasticSearchTagMapItem(l level.Level, dynamicIndexFunc ElasticSearchDynamicIndexFunc) ElasticSearchTagMapItem {
-	return ElasticSearchTagMapItem{
-		DynamicIndexFunc: dynamicIndexFunc,
-		Level:            l,
-	}
-}
-
-// ElasticSearchWithTagMap is a built-in `output` - named `ElasticSearch`,
-// that writes to ElasticSearch. It allows to define a map of tags and indexes.
-// The index name is a function which is evaluated at the insertion time.
-//
-// NOTE: By default, data is JSON-formatted.
-// NOTE: `DocumentID` is automatically generated.
-// NOTE: It's the caller's responsibility to create the index, define its
-// mapping, and settings.
-func ElasticSearchWithTagMap(
-	tagMap ElasticSearchTagMap,
-	esConfig ElasticSearchConfig,
-	processors ...processor.IProcessor,
-) []IOutput {
-	outputs := make([]IOutput, 0, len(tagMap))
-
-	for tag, elasticSearchTagMapItem := range tagMap {
-		outputs = append(outputs, ElasticSearchWithDynamicIndex(
-			elasticSearchTagMapItem.DynamicIndexFunc,
-			esConfig,
-			elasticSearchTagMapItem.Level,
-			append(processors, processor.PrintOnlyIfTagged(tag))...,
-		))
-	}
-
-	return outputs
-}
-
-// ElasticSearchWithDynamicIndex is a built-in `output` - named `ElasticSearch`,
-// that writes to ElasticSearch. It allows to define a function that returns the
-// index name to be used, evaluated at the insertion time.
-//
-// NOTE: By default, data is JSON-formatted.
-// NOTE: `DocumentID` is automatically generated.
-// NOTE: It's the caller's responsibility to create the index, define its
-// mapping, and settings.
-func ElasticSearchWithDynamicIndex(
-	dynamicIndexFunc ElasticSearchDynamicIndexFunc,
-	esConfig ElasticSearchConfig,
-	maxLevel level.Level,
-	processors ...processor.IProcessor,
-) IOutput {
-	o := New("ElasticSearch",
-		maxLevel,
-		elasticsearch.NewWithDynamicIndex(dynamicIndexFunc, esConfig),
-		processors...,
-	).SetFormatter(formatter.JSON())
-
-	return o
 }